+++ src/symbols.c
@@ -14,7 +14,7 @@ static size_t hash_str(const char *str)
 	size_t hash = 5381;
 
 	while (*str) {
-		hash = hash * 33 ^ *str;
+		hash = hash * 33 ^ (size_t)*str;
 		str++;
 	}
 
@@ -24,7 +24,7 @@ static size_t hash_str(const char *str)
 struct symbol_t *get_symbol(const char *name)
 {
 	const size_t hash = hash_str(name);
-	for (int i = 0; i < symbols[hash].count; i++)
+	for (size_t i = 0; i < symbols[hash].count; i++)
 		if (!strcmp(name, symbols[hash].data[i].name))
 			return &symbols[hash].data[i];
 	return NULL;
@@ -33,7 +33,7 @@ struct symbol_t *get_symbol(const char *name)
 struct symbol_t *create_symbol(const char *name, enum symbol_types_e type)
 {
 	const size_t hash = hash_str(name);
-	const int index = symbols[hash].count;
+	const size_t index = symbols[hash].count;
 
 	symbols[hash].count++;
 	symbols[hash].data =
@@ -57,7 +57,7 @@ size_t calc_symtab_str_buf_size(void)
 {
 	size_t sz = 1;
 	for (size_t hash = 0; hash < SYMBOLMAP_ENTRIES; hash++)
-		for (int index = 0; index < symbols[hash].count; index++)
+		for (size_t index = 0; index < symbols[hash].count; index++)
 			sz += symbols[hash].data[index].name_sz;
 	return sz;
 }
@@ -68,7 +68,7 @@ char *create_symtab_str_buf(size_t sz)
 	buf[0] = '\0';
 	size_t offset = 1;
 	for (size_t hash = 0; hash < SYMBOLMAP_ENTRIES; hash++) {
-		for (int index = 0; index < symbols[hash].count; index++) {
+		for (size_t index = 0; index < symbols[hash].count; index++) {
 			const size_t name_sz =
 				symbols[hash].data[index].name_sz;
 			memcpy(buf + offset, symbols[hash].data[index].name,
@@ -82,7 +82,7 @@ char *create_symtab_str_buf(size_t sz)
 void free_labels(void)
 {
 	for (size_t hash = 0; hash < SYMBOLMAP_ENTRIES; hash++) {
-		for (int index = 0; index < symbols[hash].count; index++)
+		for (size_t index = 0; index < symbols[hash].count; index++)
 			free(symbols[hash].data[index].name);
 		free(symbols[hash].data);
 	}
